+++ src/quark_shower.f
@@ -27,7 +27,7 @@ C...Commonblocks.
       DOUBLE PRECISION PA(6,100)
       INTEGER flav1,flav2
       DOUBLE PRECISION px,py,pz1,pz2,TAUF,minQ
-      INTEGER seed
+      INTEGER(kind=8) seed
       
       COMMON/PYJETS/N,NPAD,K(4000,5),P(4000,5),V(4000,5)
       COMMON/PYDAT1/MSTU(200),PARU(200),MSTJ(200),PARJ(200)
